+++ demos/basic-VEllipseSector/index.js
@@ -39,58 +39,42 @@
 	    var ellipse = new VEllipse( center, new Vertex(center.x+radiusH, center.y+radiusV) );
 
 	    // Now create a sector from the circle
-	    var startAngle    = 12 / 180 * Math.PI; // Math.PI*0.3; // in radians
-	    var endAngle      = 89 / 180 * Math.PI; // Math.PI*1.6;
-	    console.log('Initial: startAngle', startAngle, 'endAngle', endAngle );
+	    var startAngle    = 12 / 180 * Math.PI;
+	    var endAngle      = 89 / 180 * Math.PI;
+	    // console.log('Initial: startAngle', startAngle, 'endAngle', endAngle );
 	    var ellipseSector = new VEllipseSector( ellipse, startAngle, endAngle );
 
-	    var startControlPoint = ellipse.vertAt( startAngle ); //  new Vertex( ellipse.center.x + radiusH,  ellipse.center.y );
-	    var endControlPoint = ellipse.vertAt( endAngle ); // new Vertex( ellipse.center.x,  ellipse.center.y + radiusV );
+	    var startControlPoint = ellipse.vertAt( startAngle );
+	    var endControlPoint = ellipse.vertAt( endAngle );
 
 	    // Now add the sector to your canvas
-	    pb.add( ellipse ); // circleSector );
+	    pb.add( ellipse );
 	    pb.add( [startControlPoint, endControlPoint] );
 
 	    var startControlLine = new Line( ellipse.center, startControlPoint );
 	    var endControlLine = new Line( ellipse.center, endControlPoint );
-
-
-	    startControlPoint.listeners.addDragListener( function(event) {
-	//	pb.redraw();
-		
-	    } );
-
 	    
 	    // https://stackoverflow.com/questions/4633177/c-how-to-wrap-a-float-to-the-interval-pi-pi
 	    var wrapMax = function( x, max ) {
-		/* integer math: `(max + x % max) % max` */
-		// return fmod(max + fmod(x, max), max);
 		return (max + (x%max)) % max;
 	    };
-
 	    var wrapMinMax = function( x, min, max ) {		
-		// return min + wrapMax(x - min, max - min);
 		return min + wrapMax( x - min, max -min );
 	    };
 
 	    pb.config.postDraw = function() {
-		// console.log('postDraw');
 		pb.draw.line( startControlLine.a, startControlLine.b, 'rgba(128,128,128,0.5)', 1.0 );
 		pb.draw.line( endControlLine.a, endControlLine.b, 'rgba(128,128,128,0.5)', 1.0 );
 
 		var newStartAngle = startControlLine.angle();
 		var newEndAngle = endControlLine.angle();
-		// console.log( "newStartAngle", newStartAngle, "newEndAngle", newEndAngle );
 
 		newStartAngle = wrapMax( newStartAngle, Math.PI*2 );
 		newEndAngle = wrapMax( newEndAngle, Math.PI*2 );
-		// if( newStartAngle < 0 ) newStartAngle = newStartAngle+Math.PI*2;
-		// if( newEndAngle < 0 ) newEndAngle = newEndAngle+Math.PI*2;
 
 		var newStartPoint = ellipse.vertAt( newStartAngle );
 		var newEndPoint = ellipse.vertAt( newEndAngle );
 
-		// console.log( "newStartPoint", newStartPoint );
 		pb.draw.diamondHandle( newStartPoint, 7, 'rgba(128,64,128,0.5)' );
 		pb.draw.diamondHandle( newEndPoint, 7, 'rgba(128,64,128,0.5)' );
 
@@ -98,14 +82,14 @@
 		var pathData =
 		    VEllipseSector.ellipseSectorUtils.describeSVGArc(
 			ellipse.center.x, ellipse.center.y,
-			ellipse.radiusH(), ellipse.radiusV(), // radiusH, radiusV,
+			ellipse.radiusH(), ellipse.radiusV(),
 			newStartAngle, newEndAngle,
 			{ moveToStart : true } );
 		pb.draw.path( pathData, 'rgba(255,0,0,0.5)', 1 );
 
 		// Draw labels (start/end)
-		pb.fill.text( "start", newStartPoint.x, newStartPoint.y  ); // , 0, 'black' );
-		pb.fill.text( "end", newEndPoint.x, newEndPoint.y ); // , 0, 'black' );
+		pb.fill.text( "start", newStartPoint.x, newStartPoint.y  );
+		pb.fill.text( "end", newEndPoint.x, newEndPoint.y );
 
 		drawCircleStuff( newStartAngle, newEndAngle );
 	    };
@@ -121,13 +105,6 @@
 	    };
 
 	    pb.redraw();
-	    
-	    // Further: add a circle sector helper to edit angles and radius manually (mouse or touch)
-	    /* var controlPointA = circleSector.circle.vertAt( circleSector.startAngle );
-	    var controlPointB = circleSector.circle.vertAt( circleSector.endAngle );
-	    new CircleSectorHelper( circleSector, controlPointA, controlPointB, pb );
-	    pb.add( [controlPointA, controlPointB] );
-	    */
 	} );
     
 })(window);