+++ src/cjs/CircleSector.js
@@ -3,7 +3,8 @@
  * @author   Ikaros Kappler
  * @date     2020-12-17
  * @modified 2021-01-20 Added UID.
- * @version  1.1.0
+ * @modified 2021-02-26 Fixed an error in the svg-arc-calculation (case angle<90deg and anti-clockwise).
+ * @version  1.1.1
  **/
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.CircleSector = void 0;
@@ -94,8 +95,17 @@ var CircleSector = /** @class */ (function () {
                 return firstHalf.concat(secondHalf);
             }
             // Boolean stored as integers (0|1).
-            var largeArcFlag = endAngle - startAngle <= Math.PI ? 0 : 1;
-            var sweepFlag = 1;
+            var diff = endAngle - startAngle;
+            var largeArcFlag;
+            var sweepFlag;
+            if (diff < 0) {
+                largeArcFlag = Math.abs(diff) < Math.PI ? 1 : 0;
+                sweepFlag = 1;
+            }
+            else {
+                largeArcFlag = Math.abs(diff) > Math.PI ? 1 : 0;
+                sweepFlag = 1;
+            }
             var pathData = [];
             if (options.moveToStart) {
                 pathData.push('M', start.x, start.y);