+++ src/cjs/VEllipse.d.ts
@@ -4,7 +4,9 @@
  * @modified 2018-12-04 Added the toSVGString function.
  * @modified 2020-03-25 Ported this class from vanilla-JS to Typescript.
  * @modified 2021-01-20 Added UID.
- * @version  1.1.0
+ * @modified 2021-02-14 Added functions `radiusH` and `radiusV`.
+ * @modified 2021-02-26 Added helper function `decribeSVGArc(...)`.
+ * @version  1.2.0
  *
  * @file VEllipse
  * @fileoverview Ellipses with a center and an x- and a y-axis (stored as a vertex).
@@ -54,7 +56,23 @@ export declare class VEllipse implements SVGSerializable {
      * @name VEllipse
      **/
     constructor(center: Vertex, axis: Vertex);
+    /**
+     * Get the non-negative horizonal radius of this ellipse.
+     *
+     * @method radiusH
+     * @instance
+     * @memberof VEllipse
+     * @return {number} The horizontal radius of this ellipse.
+     */
     radiusH(): number;
+    /**
+     * Get the non-negative vertical radius of this ellipse.
+     *
+     * @method radiusV
+     * @instance
+     * @memberof VEllipse
+     * @return {number} The vertical radius of this ellipse.
+     */
     radiusV(): number;
     vertAt(angle: number): Vertex;
     /**
@@ -67,7 +85,22 @@ export declare class VEllipse implements SVGSerializable {
     toSVGString(options: {
         className?: string;
     }): string;
+    /**
+     * A static collection of ellipse-related helper functions.
+     * @static
+     */
     static utils: {
+        /**
+         * Calculate a particular point on the outline of the given ellipse (center plus two radii plus angle).
+         *
+         * @name polarToCartesian
+         * @param {number} centerX - The x coordinate of the elliptic center.
+         * @param {number} centerY - The y coordinate of the elliptic center.
+         * @param {number} radiusH - The horizontal radius of the ellipse.
+         * @param {number} radiusV - The vertical radius of the ellipse.
+         * @param {number} angle - The angle (in radians) to get the desired outline point for.
+         * @reutn {XYCoords} The outlont point in absolute x-y-coordinates.
+         */
         polarToCartesian: (centerX: number, centerY: number, radiusH: number, radiusV: number, angle: number) => XYCoords;
     };
 }