+++ src/ts/CircleSector.ts
@@ -2,7 +2,8 @@
  * @author   Ikaros Kappler
  * @date     2020-12-17
  * @modified 2021-01-20 Added UID.
- * @version  1.1.0
+ * @modified 2021-02-26 Fixed an error in the svg-arc-calculation (case angle<90deg and anti-clockwise).
+ * @version  1.1.1
  **/
 
 
@@ -148,8 +149,17 @@ export class CircleSector implements SVGSerializable {
 	    }
 
 	    // Boolean stored as integers (0|1).
-	    const largeArcFlag : number = endAngle - startAngle <= Math.PI ? 0 : 1;
-	    const sweepFlag : number = 1;
+	    const diff : number = endAngle-startAngle;
+	    var largeArcFlag : number;
+	    var sweepFlag : number;
+	    if( diff < 0 ) {
+		largeArcFlag = Math.abs(diff) < Math.PI ? 1 : 0;
+		sweepFlag = 1;
+	    } else { 
+		largeArcFlag = Math.abs(diff) > Math.PI ? 1 : 0;
+		sweepFlag = 1;
+	    }
+	    
 	    const pathData = [];
 	    if( options.moveToStart ) {
 		pathData.push('M', start.x, start.y );