+++ src/ts/module.ts
@@ -1,5 +1,9 @@
-// ? https://www.pluralsight.com/guides/react-typescript-module-create
-// ? https://blog.logrocket.com/publishing-node-modules-typescript-es-modules/
+// With the help of
+//    https://www.pluralsight.com/guides/react-typescript-module-create
+// and
+//    https://blog.logrocket.com/publishing-node-modules-typescript-es-modules/
+// 
+// @date 2021-02
 
 
 import { BezierPath } from "./BezierPath";
@@ -23,6 +27,7 @@ import { Triangle } from "./Triangle";
 import { UIDGenerator } from "./UIDGenerator";
 import { Vector } from "./Vector";
 import { VEllipse } from "./VEllipse";
+import { VEllipseSector } from "./VEllipseSector";
 import { Vertex } from "./Vertex";
 import { VertexAttr } from "./VertexAttr";
 import { VertexListeners } from "./VertexListeners";
@@ -51,6 +56,7 @@ export {
     UIDGenerator,
     Vector,
     VEllipse,
+    VEllipseSector,
     Vertex,
     VertexAttr,
     VertexListeners,
@@ -58,74 +64,3 @@ export {
     drawutilssvg
 };
 export default PlotBoilerplate;
-
-
-// const parsePath = require('./parse-path');
-// const path2dPolyfill = require('./path2d-polyfill');
-
-// if (typeof window !== 'undefined') {
-//  path2dPolyfill(window);
-//}
-
-// module.exports = {
-//  path2dPolyfill,
-//  parsePath,
-//};
-
-/*
-const BezierPath = require( "./BezierPath" );
-const Bounds = require( "./Bounds" );
-const Circle = require( "./Circle" );
-const CircleSector = require( "./CircleSector" );
-const CubicBezierCurve = require( "./CubicBezierCurve" );
-const drawutils = require( "./draw" );
-const drawutilsgl = require( "./drawgl" );
-const geomutils = require( "./geomutils" );
-const Grid = require( "./Grid" );
-// import * as interf from "./interfaces";
-const KeyHandler = require( "./KeyHandler" );
-const Line = require( "./Line" );
-const MouseHandler = require( "./MouseHandler" );
-const PBImage = require( "./PBImage" );
-const PlotBoilerplate = require( "./PlotBoilerplate" );
-const Polygon  = require( "./Polygon" );
-const SVGBuilder = require( "./SVGBuilder" );
-const Triangle = require( "./Triangle" );
-const UIDGenerator = require( "./UIDGenerator" );
-const Vector  = require( "./Vector" );
-const VEllipse = require( "./VEllipse" );
-const Vertex = require( "./Vertex" );
-const VertexAttr = require( "./VertexAttr" );
-const VertexListeners = require( "./VertexListeners" );
-const VertTuple = require( "./VertTuple" );
-const drawutilssvg = require( "./utils/helpers/drawutilssvg" );
-
-module.exports = {
-    BezierPath,
-    Bounds,
-    Circle,
-    CircleSector,
-    CubicBezierCurve,
-    drawutils,
-    drawutilsgl,
-    geomutils,
-    Grid,
-    // ...interf,
-    KeyHandler,
-    Line,
-    MouseHandler,
-    PBImage,
-    PlotBoilerplate,
-    Polygon,
-    SVGBuilder,
-    Triangle,
-    UIDGenerator,
-    Vector,
-    VEllipse,
-    Vertex,
-    VertexAttr,
-    VertexListeners,
-    VertTuple,
-    drawutilssvg
-};
-*/